TST: refactor probe classes to separate module
