+++ pyfibre/model/objects/tests/test_base_segment.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from pyfibre.tests.probe_classes import ProbeSegment
+from pyfibre.tests.probe_classes.objects import ProbeSegment
 from pyfibre.tests.pyfibre_test_case import PyFibreTestCase
