+++ pyfibre/model/tools/tests/test_convertors.py
@@ -5,7 +5,7 @@ from pyfibre.model.tools.convertors import (
     networks_to_binary, stack_to_binary, stack_to_regions,
     regions_to_stack)
 from pyfibre.tests.pyfibre_test_case import PyFibreTestCase
-from pyfibre.tests.probe_classes import (
+from pyfibre.tests.probe_classes.utilities import (
     generate_image, generate_probe_graph
 )
